lib/light: add readColorAttributes
