+++ lib/light.js
@@ -1,9 +1,46 @@
+const utils = require('./utils');
+
 async function readColorCapabilities(endpoint) {
     await endpoint.read('lightingColorCtrl', ['colorCapabilities']);
 }
 
 async function readColorTempMinMax(endpoint) {
     await endpoint.read('lightingColorCtrl', ['colorTempPhysicalMin', 'colorTempPhysicalMax']);
+}
+
+function readColorAttributes(entity, meta, additionalAttributes=[]) {
+    /**
+      * Not all bulbs suport the same features, we need to take care we read what is supported.
+      * `supportsHueAndSaturation` indicates support for currentHue and currentSaturation
+      * `enhancedHue` indicates support for enhancedCurrentHue
+      *
+      * e.g. IKEA Tådfri LED1624G9 only supports XY (https://github.com/Koenkk/zigbee-herdsman-converters/issues/1340)
+      *
+      * Additionally when we get a get payload, only request the fields included.
+     */
+    const attributes = ['colorMode'];
+
+    if (!meta.message.color || (typeof meta.message.color === 'object' && meta.message.color.hasOwnProperty('x'))) {
+        attributes.push('currentX');
+    }
+    if (!meta.message.color || (typeof meta.message.color === 'object' && meta.message.color.hasOwnProperty('y'))) {
+        attributes.push('currentY');
+    }
+
+    if (utils.getMetaValue(entity, meta.mapped, 'supportsHueAndSaturation', 'allEqual', true)) {
+        if (!meta.message.color || (typeof meta.message.color === 'object' && meta.message.color.hasOwnProperty('hue'))) {
+            if (utils.getMetaValue(entity, meta.mapped, 'enhancedHue', 'allEqual', true)) {
+                attributes.push('enhancedCurrentHue');
+            } else {
+                attributes.push('currentHue');
+            }
+        }
+        if (!meta.message.color || (typeof meta.message.color === 'object' && meta.message.color.hasOwnProperty('saturation'))) {
+            attributes.push('currentSaturation');
+        }
+    }
+
+    return [...attributes, ...additionalAttributes];
 }
 
 function findColorTempRange(entity, logger) {
@@ -66,6 +103,7 @@ async function configure(device, coordinatorEndpoint, logger, readColorTempMinMa
 module.exports = {
     readColorCapabilities,
     readColorTempMinMax,
+    readColorAttributes,
     findColorTempRange,
     clampColorTemp,
     configure,
